dom/datastore/tests/test_changes.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for DataStore - basic operation on a readonly db</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
  <script type="application/javascript;version=1.7">

  var gHostedManifestURL = 'http://test/tests/dom/datastore/tests/file_app.sjs?testToken=file_changes.html';
  var gHostedManifestURL2 = 'http://example.com/tests/dom/datastore/tests/file_app.sjs?testToken=file_changes2.html&template=file_app2.template.webapp';
  var gApps = [];
  var gApp2Events = 0;
  var gStore;

  function cbError() {
    ok(false, "Error callback invoked");
    finish();
  }

  function installApp(aApp) {
    var request = navigator.mozApps.install(aApp);
    request.onerror = cbError;
    request.onsuccess = function() {
      gApps.push(request.result);
      runTest();
    }
  }

  function uninstallApps() {
    if (!gApps.length) {
      ok(true, "All done!");
      runTest();
      return;
    }

    var app = gApps.pop();
    var request = navigator.mozApps.mgmt.uninstall(app);
    request.onerror = cbError;
    request.onsuccess = uninstallApps;
  }

  function setupApp2() {
    var ifr = document.createElement('iframe');
    ifr.setAttribute('mozbrowser', 'true');
    ifr.setAttribute('mozapp', gApps[1].manifestURL);
    ifr.setAttribute('src', gApps[1].manifest.launch_path);
    var domParent = document.getElementById('content');

    // Set us up to listen for messages from the app.
    var listener = function(e) {
      var message = e.detail.message;
      if (/^OK/.exec(message)) {
        ok(true, "Message from app: " + message);
      } else if (/KO/.exec(message)) {
        ok(false, "Message from app: " + message);
      } else if (/READY/.exec(message)) {
        ok(true, "App2 ready");
        runTest();
      } else if (/DONE/.exec(message)) {
        ok(true, "Messaging from app complete");
        ifr.removeEventListener('mozbrowsershowmodalprompt', listener);
        domParent.removeChild(ifr);
        gApp2Events++;
      }
    }

    // This event is triggered when the app calls "alert".
    ifr.addEventListener('mozbrowsershowmodalprompt', listener, false);
    domParent.appendChild(ifr);
  }

  function testApp1() {
    var ifr = document.createElement('iframe');
    ifr.setAttribute('mozbrowser', 'true');
    ifr.setAttribute('mozapp', gApps[0].manifestURL);
    ifr.setAttribute('src', gApps[0].manifest.launch_path);
    var domParent = document.getElementById('content');

    // Set us up to listen for messages from the app.
    var listener = function(e) {
      var message = e.detail.message;
      if (/^OK/.exec(message)) {
        ok(true, "Message from app: " + message);
      } else if (/KO/.exec(message)) {
        ok(false, "Message from app: " + message);
      } else if (/DONE/.exec(message)) {
        ok(true, "Messaging from app complete");
        ifr.removeEventListener('mozbrowsershowmodalprompt', listener);
        domParent.removeChild(ifr);
        runTest();
      }
    }

    // This event is triggered when the app calls "alert".
    ifr.addEventListener('mozbrowsershowmodalprompt', listener, false);
    domParent.appendChild(ifr);
  }

  function checkApp2() {
    ok(gApp2Events, "App2 received events");
    runTest();
  }

  var tests = [
    // Permissions
    function() {
      SpecialPowers.pushPermissions(
        [{ "type": "browser", "allow": 1, "context": document },
         { "type": "embed-apps", "allow": 1, "context": document },
         { "type": "webapps-manage", "allow": 1, "context": document }], runTest);
    },

    // Preferences
    function() {
      SpecialPowers.pushPrefEnv({"set": [["dom.promise.enabled", true]]}, runTest);
    },

    function() {
      SpecialPowers.pushPrefEnv({"set": [["dom.datastore.enabled", true]]}, runTest);
    },

    function() {
      SpecialPowers.pushPrefEnv({"set": [["geo.testing.ignore_ipc_principal", true]]}, runTest);
    },

    // Enabling mozBrowser
    function() {
      SpecialPowers.setAllAppsLaunchable(true);
      SpecialPowers.pushPrefEnv({"set": [["dom.mozBrowserFramesEnabled", true]]}, runTest);
    },

    // No confirmation needed when an app is installed
    function() {
      SpecialPowers.autoConfirmAppInstall(runTest);
    },

    // Installing the app1
    function() { installApp(gHostedManifestURL); },

    // Installing the app2
    function() { installApp(gHostedManifestURL2); },

    // Setup app2 for receving events
    setupApp2,

    // Run tests in app
    testApp1,

    // Check app2
    checkApp2,

    // Uninstall the apps
    uninstallApps,
  ];

  function runTest() {
    if (!tests.length) {
      finish();
      return;
    }

    var test = tests.shift();
    test();
  }

  function finish() {
    SimpleTest.finish();
  }

  if (SpecialPowers.isMainProcess()) {
    SpecialPowers.Cu.import("resource://gre/modules/DataStoreChangeNotifier.jsm");
  }

  SimpleTest.waitForExplicitFinish();
  runTest();
  </script>
</pre>
</body>
</html>
